requirements.txt
mediapipe
opencv-python
keyboard
pyttsx3
pygame
pyserial